# back-end/docs/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
    "/api/users/login": {
      "post": {
        "operationId": "UsersController_login",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "Login with email and password",
        "tags": [
          "Users"
        ]
      }
    },
    "/api/users": {
      "get": {
        "operationId": "UsersController_findAll",
        "parameters": [
          {
            "name": "role",
            "required": false,
            "in": "query",
            "description": "Filter by role (client, worker, expert, superuser)",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Get all users",
        "tags": [
          "Users"
        ]
      },
      "post": {
        "operationId": "UsersController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "Register a new user",
        "tags": [
          "Users"
        ]
      }
    },
    "/api/users/{id}": {
      "get": {
        "operationId": "UsersController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Get user by ID",
        "tags": [
          "Users"
        ]
      },
      "patch": {
        "operationId": "UsersController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "role",
            "in": "header",
            "description": "User role required",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Update user profile",
        "tags": [
          "Users"
        ]
      },
      "delete": {
        "operationId": "UsersController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "role",
            "in": "header",
            "description": "User role required",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Delete a user (superuser only)",
        "tags": [
          "Users"
        ]
      }
    },
    "/api/users/{id}/wallet/add": {
      "post": {
        "operationId": "UsersController_addToWallet",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "role",
            "in": "header",
            "description": "User role required",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WalletDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "Add funds to user wallet",
        "tags": [
          "Users"
        ]
      }
    },
    "/api/users/{id}/wallet/deduct": {
      "post": {
        "operationId": "UsersController_deductFromWallet",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "role",
            "in": "header",
            "description": "User role required",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WalletDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "Deduct funds from user wallet",
        "tags": [
          "Users"
        ]
      }
    },
    "/api/tasks": {
      "get": {
        "operationId": "TasksController_findAll",
        "parameters": [
          {
            "name": "clientId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "workerId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Get all tasks (supports filtering)",
        "tags": [
          "Tasks"
        ]
      },
      "post": {
        "operationId": "TasksController_create",
        "parameters": [
          {
            "name": "role",
            "in": "header",
            "description": "User role required",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTaskDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "Create a new task",
        "tags": [
          "Tasks"
        ]
      }
    },
    "/api/tasks/{id}": {
      "get": {
        "operationId": "TasksController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Get task by ID",
        "tags": [
          "Tasks"
        ]
      },
      "patch": {
        "operationId": "TasksController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "role",
            "in": "header",
            "description": "User role required",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTaskDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Update a task",
        "tags": [
          "Tasks"
        ]
      },
      "delete": {
        "operationId": "TasksController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "role",
            "in": "header",
            "description": "User role required",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Delete a task",
        "tags": [
          "Tasks"
        ]
      }
    },
...
    "/api/transactions": {
      "get": {
        "operationId": "TransactionsController_findAll",
        "parameters": [
          {
            "name": "userId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Get all transactions (supports ?userId= filter)",
        "tags": [
          "Transactions"
        ]
      },
      "post": {
        "operationId": "TransactionsController_create",
        "parameters": [
          {
            "name": "role",
            "in": "header",
            "description": "User role required",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTransactionDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "Create a transaction",
        "tags": [
          "Transactions"
        ]
      }
    },
...
  },
  "info": {
    "title": "Lannent API",
    "description": "Complete REST API for the Lannent freelance platform. Supports Users, Tasks, Milestones, Proposals, Audit Requests, Audit Reports, Disputes, Transactions, Expert Applications, and Notifications. Uses in-memory storage with seed data. Include \"role\" and \"user-id\" headers for RBAC.",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
      "LoginDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "client@gmail.com"
          },
          "password": {
            "type": "string",
            "example": "Password@123"
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
      "CreateUserDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "John Doe"
          },
          "email": {
            "type": "string",
            "example": "john@gmail.com"
          },
          "password": {
            "type": "string",
            "example": "Password@123"
          },
          "role": {
            "type": "string",
            "example": "client",
            "enum": [
              "client",
              "worker",
              "expert",
              "superuser"
            ]
          },
          "avatar": {
            "type": "string",
            "example": "JD"
          },
          "avatarColor": {
            "type": "string",
            "example": "linear-gradient(135deg,#6366f1,#4f46e5)"
          },
          "company": {
            "type": "string",
            "example": "TechCorp"
          },
          "location": {
            "type": "string",
            "example": "New York, NY"
          },
          "skills": {
            "example": [
              "React",
              "Node.js"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "specialization": {
            "type": "string",
            "example": "Full-Stack & Security"
          }
        },
        "required": [
          "name",
          "email",
          "password",
          "role"
        ]
      },
      "UpdateUserDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "John Doe"
          },
          "email": {
            "type": "string",
            "example": "john@gmail.com"
          },
          "password": {
            "type": "string",
            "example": "Password@123"
          },
          "role": {
            "type": "string",
            "example": "client",
            "enum": [
              "client",
              "worker",
              "expert",
              "superuser"
            ]
          },
          "avatar": {
            "type": "string",
            "example": "JD"
          },
          "avatarColor": {
            "type": "string",
            "example": "linear-gradient(135deg,#6366f1,#4f46e5)"
          },
          "company": {
            "type": "string",
            "example": "TechCorp"
          },
          "location": {
            "type": "string",
            "example": "New York, NY"
          },
          "skills": {
            "example": [
              "React",
              "Node.js"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "specialization": {
            "type": "string",
            "example": "Full-Stack & Security"
          },
          "status": {
            "type": "string",
            "example": "active",
            "enum": [
              "active",
              "suspended"
            ]
          },
          "walletBalance": {
            "type": "number",
            "example": 10000
          },
          "rating": {
            "type": "number",
            "example": 4.8
          },
          "completedProjects": {
            "type": "number",
            "example": 25
          },
          "reviewsDone": {
            "type": "number",
            "example": 50
          }
        }
      },
      "WalletDto": {
        "type": "object",
        "properties": {
          "amount": {
            "type": "number",
            "example": 500,
            "description": "Amount to add or deduct"
          }
        },
        "required": [
          "amount"
        ]
      },
      "CreateTaskDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "example": "E-commerce Website Redesign"
          },
          "description": {
            "type": "string",
            "example": "Need a complete redesign of our platform"
          },
          "category": {
            "type": "string",
            "example": "Web Development"
          },
          "budget": {
            "type": "number",
            "example": 2500
          },
          "currency": {
            "type": "string",
            "example": "USD"
          },
          "deadline": {
            "type": "string",
            "example": "2026-04-15"
          },
          "skills": {
            "example": [
              "React",
              "Figma"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "clientId": {
            "type": "string",
            "example": "u1"
          },
          "auditEnabled": {
            "type": "boolean",
            "example": false
          }
        },
        "required": [
          "title",
          "description",
          "category",
          "budget",
          "clientId"
        ]
      },
      "UpdateTaskDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "example": "E-commerce Website Redesign"
          },
          "description": {
            "type": "string",
            "example": "Need a complete redesign of our platform"
          },
          "category": {
            "type": "string",
            "example": "Web Development"
          },
          "budget": {
            "type": "number",
            "example": 2500
          },
          "currency": {
            "type": "string",
            "example": "USD"
          },
          "deadline": {
            "type": "string",
            "example": "2026-04-15"
          },
          "skills": {
            "example": [
              "React",
              "Figma"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "clientId": {
            "type": "string",
            "example": "u1"
          },
          "auditEnabled": {
            "type": "boolean",
            "example": false
          },
          "status": {
            "type": "string",
            "example": "in-progress",
            "enum": [
              "open",
              "in-progress",
              "under-review",
              "completed",
              "cancelled"
            ]
          },
          "workerId": {
            "type": "string",
            "example": "u5"
          },
          "progress": {
            "type": "number",
            "example": 75
          }
        }
      },
...
      "CreateTransactionDto": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "example": "escrow-lock",
            "enum": [
              "escrow-lock",
              "milestone-release",
              "deposit",
              "refund",
              "dispute-release"
            ]
          },
          "amount": {
            "type": "number",
            "example": 2500
          },
          "fromId": {
            "type": "string",
            "example": "u1"
          },
          "toId": {
            "type": "string",
            "example": "escrow"
          },
          "taskId": {
            "type": "string",
            "example": "t1"
          },
          "milestoneId": {
            "type": "string",
            "example": "m1"
          },
          "description": {
            "type": "string",
            "example": "Escrow funded for project"
          },
          "status": {
            "type": "string",
            "example": "completed"
          }
        },
        "required": [
          "type",
          "amount",
          "fromId",
          "toId"
        ]
      },
...
    }
  }
}